Added German translation to Sidebar Menu
I've added the appropriate German translation to the sidebar menu. I hope everything fits. If not, it may be needed to adjust the width of it, as many sentences are longer in German than in English. Now the complete map can be used in German.  

I'm a native German speaker who used the German version of the game to make sure everything is synced with it - I normally have set my PS4 to English. I'm fluent in both languages.