Load weekly set earlier && little cleanup
I know it not solves the problem but reduce to minimum errors when
nothing load on the map due to not loaded weekly set.
The error is related to asynchronous data retrieval,
none of the function checks if everything has been loaded correctly and
then functions works on undefined data.