Make item names consistent with the game.
Not every wild flower is a "flower" (creek plum is a berry), nor is every coin a "coin" (a token isn't a coin), so just use the in-game names for all of the items. Only items that aren't 100% the same as the game are the cards, which should all have "Tarot Card" pasted after them (like in the game), but that's kind of useless.