Sunsetting cycles display by default.
Yes, for real. The concept of "cycles" is confusing more and more people every day. The default display should be false, with a way to enable it. People should just see the icons on the map and go to them, since they are correct 99% of the time. All the parts are still there for people that use cycles, though I am honestly less and less sure by the day that people actually do.

People get stuck in the concept of "day 1 is Monday, day 2 is Tuesday" and so on, and they confuse the numbers of the days with cycles. People think there are only 3 cycles. People think they need to manually configure the cycles. The list of people getting confused by the concept goes on and on.

Let's keep an eye on how this goes for now. This is easy to change if need be.

This commit does the following:
- Hide cycles by default;
- Fix up Nazar date getting stuck on a language;
- Use "numeric" instead of "2-digit", it makes more sense in a greater number of languages, like Asian ones, and some other ones;
- Reorganize the menu slightly;
- Slightly tweak the modals.